Accelerated-NEAT implement. Quite different from what I was planning initially. The weight mutations of a champion are repeated in its child in the next generation.
After this, try out Leaky-ReLu and the possibilities mentioned in the Original implementation's comments as 'possibilities'.

!!THE CODE IS YET TO BE TESTED AND THE HYPER-PARAMETERS ARE TO BE OPTIMESED!!
To be considered a success, the total number of runs before a solution is found should be at most half of the original. The population must be at most 10% of the original. The number of generations required would be greater.